+++ backend/src/controllers/songsControllers.js
@@ -1,29 +1,40 @@
 import { check, validationResult } from "express-validator";
 import { Song } from "../models/song.js";
 
+//GET all songs
 export const getSongs = async (req, res) => {
   res.set("Access-Control-Allow-Origin", "http://localhost:3000");
   const songs = await Song.find();
   res.status(200).send(songs);
 };
 
+//GET with ID
 export const getSongById = async (req, res) => {
-  let song = await Song.findById(req.params.id);
-  res.status(200).send(song);
+  //check if ID has 24 characters
+  if (req.params.id.length != 24) {
+    res.status(400).send("ID must have 24 characters!");
+  }else {
+    let song = await Song.findById(req.params.id);
+    res.status(200).send(song);
+  }
 };
 
+//GET with specific title
 export const getSongByTitle = async (req, res) => {
   let song = await Song.find({ title: req.query.title });
   res.status(200).send(song);
 };
 
+//POST for new song
 export const addSong = async (req, res) => {
+  //check if data contains all attributes
   const error = validationResult(req);
 
   if (!error.isEmpty()) {
     return res.status(400).json({ error: error.array() });
   }
 
+  //create new song
   const song = new Song({
     title: req.body.title,
     year: req.body.year,
@@ -35,23 +46,32 @@ export const addSong = async (req, res) => {
   song.save(song).then((todo) => res.status(201).send(todo));
 };
 
+//PATCH for song by ID
 export const patchSong = async (req, res) => {
-  const error = validationResult(req);
+  //check if ID has 24 characters
+  if (req.params.id.length != 24) {
+    res.status(400).send("ID must have 24 characters!");
+  //check if body has no ID else answer with 400
+  }else if(req.body._id) {
+    res.status(400).send("Cannot change ID!");
+  }else {
+    let response = await Song.findByIdAndUpdate(req.params.id, req.body, {
+      new: false,
+    });
 
-  if (!error.isEmpty()) {
-    return res.status(400).json({ error: error.array() });
+    res.status(200).send(response);
   }
-
-  let response = await Song.findByIdAndUpdate(req.params.id, req.body, {
-    new: false,
-  });
-
-  res.status(200).send(response);
 };
 
+//DELETE song by ID
 export const deleteSong = async (req, res) => {
-  let response = await Song.findByIdAndDelete(req.params.id);
-  res.status(200).send(response);
+  //check if ID has 24 characters
+  if (req.params.id.length != 24) {
+    res.status(404).send("ID must have 24 characters!");
+  }else {
+    let response = await Song.findByIdAndDelete(req.params.id);
+    res.status(200).send(response);
+  }
 };
 
 // attached as second param in a route
